+++ src/parser/formalargs.rs
@@ -1,6 +1,6 @@
-use formalargs::{CallArgs, FormalArgs};
 use parser::util::{ignore_comments, name, opt_spacelike};
 use parser::value::{extended_literal, space_list};
+use sass::{CallArgs, FormalArgs};
 use sass::Value;
 
 named!(pub formal_args<FormalArgs>,
