+++ src/sass/mod.rs
@@ -1,5 +1,9 @@
 mod item;
 mod value;
+mod formal_args;
+mod call_args;
 
+pub use self::call_args::CallArgs;
+pub use self::formal_args::FormalArgs;
 pub use self::item::Item;
 pub use self::value::Value;
