+++ src/sass/value.rs
@@ -1,8 +1,8 @@
 use error::Error;
-use formalargs::CallArgs;
 use functions::get_builtin_function;
 use num_rational::Rational;
 use num_traits::{One, Signed, Zero};
+use sass::CallArgs;
 use std::fmt;
 use value::{ListSeparator, Operator, Quotes, Unit, rgb_to_name};
 use variablescope::Scope;
